Atualizando estrutura para prova de conceito.

+++ server/SportPoint.Server.Dao.Collection/JogadorDaoCollection.cs
@@ -61,7 +61,7 @@ namespace SportPoint.Server.Dao.Collection
             throw new NotImplementedException();
         }
 
-        public int Count()
+        public long Count()
         {
             throw new NotImplementedException();
         }
